requirements.txt
discord.py
pymongo
pymongo[srv]
python-dotenv
requests